- Add OpenGL to premake.lua, refactor per platform external linking libraries. - Add Editor icons via the new TextureAtlas system. - Now allow node deletion from the outliner. Also display entities like "lights" in the outliner. - Constraint some shortcut to the viewport in hovered state only. - Rewrite the FileBrowser component (Still an issue: when clicking the last child of a node, it closes the parent, wtf?). ImGui related. - Implement the new icons in the outliner. Also stylizing it better. - A lot of changes in the properties editor : Now display mesh properties & material properties. Also per light infos / oiptions. It is possible to switch the light type at runtime but not the material (soon implemented). Texture update at runtime is also implemented. - Improve the tools panel for creating the entities. - refactor the viewport and moving elements to their respective locations of external class (like the gizmo). Also started to implemented viewport split to view different frame buffer attachments. - Fix a lot of bugs in AssimpImporter and also reformat the vertex data to be conform to the new buffers structures. - Add a new Gizmo class to manipulate the entities in the viewport. (ImGuizmo); - Changes class Node, StaticMesh & Material to be  constructed via std::shared_ptr to avoid memory leaks. - Improve the selection system for the entities via the outliner. - Finally implemented the basic shapes for the physics engine (Plane, Box & Sphere). - Rewrote almost every opengl buffers to be simple to use. - Fix various bugs in the camera classes and implement the panning for the Third person camera. - Add the new SoundManager class to be in pair with Opengl and the audio system. - Separating the concerns between the forward and deferred rendering to allow better performance with multiples lights later. - A  new utility classes : treeNode, lerp, randomPointInCircle and also fileDialog (file picker for windows only). - Improves the geometry class, now provide better default vertex data for each shapes. Also fix various bugs and improve the api usage. - Updated the StaticMesh calss, new methods:  setupInstances ,updateInstance, calculateTangents. rewrote the setupBuffer Mesh. Updated the header to be readable by the PropertiesEditor. - Prepaer the lighting system to carry their own shadow generator per light. Also add the missing intensity parameter xD. - Improved the shader system - Fix a bug with texture instancing creating a memory leak. Now correctly freeing the image. - Add new parameters to the default PhongMaterial preset. - Add various tests and debugs in the ForwardRenderer. (For instancing for example) - attach lights to the scene system. - improved the default vertex shader - Add new tests in the sandbox project main file. - Add new assets for testing the engine. - Add new texturing for the landscape generator. Now supports splatmaps for each lights. - add new WavLoader for the audio system - Fix various bugs and add minor features.

+++ Razor/src/Razor/Geometry/Geometry.h
@@ -3,7 +3,8 @@
 #include "StaticMesh.h"
 #include "Razor/Maths/Maths.h"
 
-namespace Razor {
+namespace Razor 
+{
 
 	class Quad : public StaticMesh
 	{
@@ -50,15 +51,15 @@ namespace Razor {
 	class Grid : public StaticMesh
 	{
 	public:
-		Grid(float divisions = 15.0f);
+		Grid(int divisions = 15);
 		~Grid();
 
-		inline void setDivisions(float divisions) { this->divisions = divisions; }
-		inline float& getDivisions() { return this->divisions; }
+		inline void setDivisions(int divisions) { this->divisions = divisions; }
+		inline int& getDivisions() { return this->divisions; }
 		void generate();
 
 	private:
-		float divisions;
+		int divisions;
 	};
 
 	class Ray : public StaticMesh
